USB Gadget: Convert from HIDL to AIDL
Convert USB Gadget HAL implementation from HIDL to AIDL.

The implementation continues to link to the libusbconfigfs that
is provided by the HIDL-based android.hardware.usb.gadget@1.1
default implementation, as the AIDL default implementation
currently does not expose the useful MonitorFfs and misc utility
helpers.  It links only to the lib and not the actual HIDL or
binder bits, so the end result is still a pure AIDL service.

Also add the implementation for the new API getUsbSpeed() defined
in the AIDL interface.

Change-Id: I3f2bd6c7c4f5904d36d2261ab0b38a06c8976d38