USB: HAL: Add correct set mode for Dual role classes
Currently the we get a queryPortStatus or a type-c uevent
it goes ahead and calls getPortStatusHelper which tries to
get PortMode requested by UsbPortManager.
However, in cases of targets using kernel PD based solution
the function getCurrentRoleHelper when tried to get the role,
it comes from  port_type instead of data_role/power_role.
This is because the kernel PD lets the device advertise as
a DRP device. So when "dual" is returned as the role, the
HAL currently returns the status as unrecognized role, to which
the UsbPortManager gets into failure path and result in the
USB preferences greyed out.

Fix this by setting the PortMode as DRP instead of NONE and
return SUCCESS instead of UNRECOGNIZED_ROLE.

Change-Id: Ic6fba49d7ec385010b5205bbe7a4609f7bd7d706