USB: HAL: Retry binding to UDC while adb disabled
Currently, the USB HAL uevent listener will attempt to write to the UDC
configfs file when adb is disabled, and user is switching between host and
device modes.  The Android USB framework will also have attempts to write
to the UDC file, however, there is a situation where these writes can
fail due to other services not being ready at the time of binding.  Add a
retry mechanism to ensure that the write to the UDC file is complete.

Change-Id: I6de0db3948176f7f05b39fafe8da4bb1492a71b7