USB: HAL: Fix PortRole mode type usecases for assigning role correctly
Commit e51d302dc9a4 (“USB: Various minor style cleanups”) reworked the
getCurrentRoleHelper function to use the appendRoleNodeHelper API for
setting the correct role. However, this change inadvertently caused the
function to read port_type instead of data_role. This misalignment led
to confusion in the HAL, as the values of source/sink for port_type
resemble those of power_role.

To resolve this issue, revert to the previous method of directly using
the filename and ensure data_role is read when port_type is present.
Also, removing the dual check since this won't be used anymore.

Change-Id: Idadd639a22caa5bb0b49b9b8edfb60725fa064f3
(cherry picked from commit 2586dd4b9239fcb146bfb840c2f311a746ed1763)